+++ CHANGELOG.rst
@@ -4,9 +4,13 @@ CHANGELOG
 Development
 -----------
 
+- [Fix] Load ``freq_min_hz`` and ``freq_max_hz`` values from metadata
+  (`MR126 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/126>`__)
+- [Fix] Strip whitespaces from beginning and end of ``catalogue_name`` before loading to database
+  (`MR126 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/126>`__)
 - [Update] Added ECSV output format for use with TOPCAT
   (`MR125 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/125>`__)
-- [Update] Move all in-memory state to the DB. And allow listing of previous upload information.
+- [Update] Move all in-memory state to the DB. And allow listing of previous upload information
   (`MR124 <https://gitlab.com/ska-telescope/sdp/ska-sdp-global-sky-model/-/merge_requests/124>`__)
 
 0.4.0